feat(gateway): extract composable engine server into the lib
Move AppState, auth, S3/key/plugin/backend handlers, and the router into a
new server module so the private SaaS crate can compose the engine. Thread
ControlPlane::authorize/record through the S3 handlers. The OSS self-host
binary is now thin and uses NoopControlPlane.
